requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
transformers
torch
chromadb
sentence-transformers
openai-whisper
pypdf
tiktoken
ffmpeg-python